chore: add .gitignore file to ignore node_modules directory
chore: add LICENSE file with MIT license

chore: add next.js configuration file for ESLint

chore: add package.json file with dependencies and peer dependencies

The changes were made to add necessary files and configurations for the project. The .gitignore file was added to ignore the node_modules directory, which contains the installed dependencies. The LICENSE file was added with the MIT license to specify the terms of use for the project. The next.js configuration file was added to configure ESLint for the project. Lastly, the package.json file was added with the necessary dependencies and peer dependencies for the project.